rustsbi
Derive Macro RustSBI

#[derive(RustSBI)]
{
    // Attributes available to this derive:
    #[rustsbi]
}

Generate RustSBI implementation for structure of each extension.

Usage

The #[derive(RustSBI)] macro provides a convenient way of building RustSBI trait implementations. To use this macro, say that we have a struct MyFence with RISC-V SBI Remote Fence extension implemented using rustsbi::Fence trait. Then, we build a struct around it, representing a whole SBI implementation including one Fence extension only; we can name it MySBI:

struct MySBI {
    fence: MyFence,
}

struct MyFence { /* fields */ }

impl rustsbi::Fence for MyFence {
    /* implementation details */
}

Here, we declared the field named fence with type MyFence. The variable name fence is special, it tells the RustSBI derive-macro that this field implements SBI Remote Fence instead of other SBI extensions. We can continue to add more fields into MySBI. For example, if we have RISC-V SBI Time extension implementation with type MyTimer, we can add it to MySBI:

struct MySBI {
    fence: MyFence,
    timer: MyTimer,
}

Don’t forget that the name timer is also a special field name. There is a detailed list after this chapter describing what special field name would the RustSBI macro identify.

It looks like we are ready to derive RustSBI macro on MySBI! Let’s try it now …

#[derive(RustSBI)]
struct MySBI {
    fence: MyFence,
    timer: MyTimer,
}

Oops! Compile failed. We’d check what happened here:

error: can't derive RustSBI: #[cfg(feature = "machine")] is needed to derive RustSBI with no extra `EnvInfo` provided; consider adding an `info` parameter to provide machine information implementing `rustsbi::EnvInfo` if RustSBI is not run on machine mode.
 --> example.rs:LL:10
   |
LL | #[derive(RustSBI)]
   |          ^^^^^^^
   |
 = note: this error originates in the derive macro `RustSBI` (in Nightly builds, run with -Z macro-backtrace for more info)

error: aborting due to previous error

The error message hints that we didn’t provide any SBI environment information implementing trait EnvInfo. By default, RustSBI is targeted to provide RISC-V supervisor environment on any hardware, focusing on hypervisor, emulator and binary translation applications. In this case, the virtualized environment should provide the supervisor with machine environment information like mvendorid, marchid and mimpid values. RustSBI could also be used on bare-metal RISC-V machines where such values would be directly accessible through CSR read operations.

If we are targeting bare-metal, we can use the RustSBI library with #[cfg(feature = "machine")] enabled by changing dependencies section in Cargo.toml file (if we are using Cargo):

[dependencies]
rustsbi = { version = "0.4.0", features = ["machine"] }

If that’s not the case, and we are writing a virtualization-targeted application, we should add a EnvInfo implementation into the structure like MySBI mentioned above, with the special field name info. We can do it like:

#[derive(RustSBI)]
struct MySBI {
    fence: MyFence,
    timer: MyTimer,
    info: MyEnvInfo,
}

struct MyEnvInfo;

impl rustsbi::EnvInfo for MyEnvInfo {
    #[inline]
    fn mvendorid(&self) -> usize { todo!("add real value here") }
    #[inline]
    fn marchid(&self) -> usize { todo!("add real value here") }
    #[inline]
    fn mimpid(&self) -> usize { todo!("add real value here") }
}

Then, when we compile our code with MySBI, we’ll find that the code now compiles successfully.

To use the derived RustSBI implementation, we note out that this structure now implements the trait RustSBI with function handle_ecall. We can pass the SBI extension, function and parameters into handle_ecall, and read the SBI call result from its return value with the type SbiRet. To illustrate this feature, we make an SBI call to read the SBI implementation ID, like:

#[derive(RustSBI)]
struct MySBI {
    /* we omit the extension fields by now */
}

fn main() {
    // Create a MySBI instance.
    let sbi = MySBI {
        /* include initial values for fields */
    };
    // Make the call. Read SBI implementation ID resides in extension Base (0x10),
    // with function id 1, and it doesn't have any parameters.
    let ret = sbi.handle_ecall(0x10, 0x1, [0; 6]);
    // Let's check the result...
    println!("SBI implementation ID for MySBI: {}", ret.value);
    assert_eq!(ret.value, 4);
}

Run the code, and we’ll find the following output in the console:

SBI implementation ID for MySBI: 4

The SBI call returns the number 4 as the SBI call result. By looking up the RISC-V SBI Specification, we can know that RustSBI have the implementation ID of 4. You have successfully made your first SBI call from a derived RustSBI implementation!

If we learn further from the RISC-V privileged software architecture, we may know more about how RISC-V SBI works on an environment to support supervisor software. RISC-V SBI implementations accept SBI calls by supervisor-level environment call caused by ecall instruction under supervisor mode. Each ecall raises a RISC-V exception which the environment must process with. The SBI environment, either bare-metal or virtually, would save context, read extension, function and parameters and call the handle_ecall function provided by RustSBI trait. Then, the function returns with an SbiRet; we read back value and error to store them into the saved context. Finally, when the context restores, the supervisor mode software (kernels, etc.) could get the SBI call result from register values.

Additionally, the macro also provides a dynamic way of building RustSBI trait implementations. It allows developers to dynamically choose which device to use in the actual RustSBI implementation. To use this feature, consider two structs namely FenceOne and FenceTwo, both implementing RISC-V SBI Remote Fence extension. We can now derive RustSBI macro on MySBI wrapping those two structs with Option type, annotated with #[rustsbi(dynamic)]: ``

#[derive(RustSBI)]
#[rustsbi(dynamic)]
struct MySBI {
    fence: Option<FenceOne>,
    rfnc: Option<FenceTwo>,
}

struct FenceOne { /* fields */ }

struct FenceTwo { /* fields */ }

// Both `FenceOne` and `FenceTwo` implements `rustsbi::Fence`.

We have declared two fields, fence and rfnc, both of which are recognized by RustSBI as potential candidates for the SBI Remote Fence extension. Dynamic RustSBI will probe both fields to determine which field it should use for the SBI calls on the Remote Fence extension.

Now we have learned basic usages of the derive-macro RustSBI. We can dive deeper and use RustSBI in real cases with ease. Congratulations!

Supported extensions

The derive macro RustSBI supports all the standard RISC-V SBI extensions this library supports. When we add extensions into SBI structure fields, special field names are identified by RustSBI derive macro. Here is a list of them:

Field names	RustSBI trait	Extension

time or timer	Timer	Timer programmer extension

ipi or spi	Ipi	S-mode Inter Processor Interrupt

fence or rfnc	Fence	Remote Fence extension

hsm	Hsm	Hart State Monitor extension

reset or srst	Reset	System Reset extension

pmu	Pmu	Performance Monitor Unit extension

console or dbcn	Console	Debug Console extension

susp	Susp	System Suspend extension

cppc	Cppc	SBI CPPC extension

nacl	Nacl	Nested Acceleration extension

sta	Sta	Steal Time Accounting extension

The EnvInfo parameter is used by RISC-V SBI Base extension which is always supported on all RISC-V SBI implementations. RustSBI provides the Base extension with additional EnvInfo by default.

Field names	RustSBI trait	Description

info or env_info	EnvInfo	Machine environment information used by Base extension

Or, if #[cfg(feature = "machine")] is enabled, RustSBI derive macro does not require additional machine environment information but reads them by RISC-V CSR operation when we don’t have any EnvInfos in the structure. This feature would only work if RustSBI runs directly on machine mode hardware. If we are targeting other environments (virtualization etc.), we should provide EnvInfo instead of using the machine feature.

Examples

This macro should be used over a struct of RISC-V SBI extension implementations. For example:

#[derive(RustSBI)]
struct MySBI {
    fence: MyFence,
    info: MyEnvInfo,
}

// Here, we assume that `MyFence` implements `rustsbi::Fence`
// and `MyEnvInfo` implements `rustsbi::EnvInfo`.

Fields indicating the same extension (SBI extension or EnvInfo) shouldn’t be included more than once.

#[derive(RustSBI)]
struct MySBI {
    fence: MyFence,
    rfnc: MyFence, // <-- Second field providing `rustsbi::Fence` implementation
    info: MyEnvInfo,
}

When using the #[rustsbi(dynamic)] attribute, it is possible to include multiple fields that indicate the same SBI extension.

#[derive(RustSBI)]
#[rustsbi(dynamic)]
struct MySBI {
    // Fields in `#[rustsbi(dynamic)]` structures are usually wrapped in `Option`s.
    fence: Option<MyFence>,
    rfnc: Option<MyFence>,
    // ^ Both the two fields are identified as `rustsbi::Fence` implementation.
    info: MyEnvInfo,
    // ^ However, environment information should only be introduced *once*.
}

// RustSBI will sequentially examine these fields, starting from the first one.
// Upon encountering the first `Option` that is `Some`, it will utilize this
// field as the implementation.
// i.e., if the first field `fence` in `MySBI` is `Some`, RustSBI uses `fence`.
// Else, if the second field `rfnc` is `Some`, RustSBI uses `rfnc`.
// Otherwise, RustSBI returns `SbiRet::not_supported`.

The struct as derive input may include generics, specifically type generics, lifetimes, constant generics and where clauses.

#[derive(RustSBI)]
struct MySBI<'a, T: rustsbi::Fence, U, const N: usize>
where
    U: rustsbi::Timer,
{
    fence: T,
    timer: U,
    info: &'a MyEnvInfo,
    _dummy: [u8; N],
}

Inner attribute #[rustsbi(skip)] informs the macro to skip a certain field when generating a RustSBI implementation.

#[derive(RustSBI)]
struct MySBI {
    console: MyConsole,
    #[rustsbi(skip)]
    fence: MyFence,
    info: MyEnvInfo,
}

// The derived `MySBI` implementation ignores the `fence: MyFence`. It can now
// be used as a conventional struct field.
// Notably, a `#[warn(unused)]` would be raised if `fence` is not further used
// by following code; `console` and `info` fields are not warned because they are
// internally used by the trait implementation derived in the RustSBI macro.

In some cases, we may manually assign fields to a certain SBI extension other than defaulting to special names defined above, and sometimes we need to provide multiple SBI extensions with one field only. By listing the extension names separated by comma in the helper attribute, we can assign one or multiple SBI extensions to a field to solve the issues above.

#[derive(RustSBI)]
struct MySBI {
    console: MyConsole,
    #[rustsbi(fence)]
    some_other_name: MyFence,
    info: MyEnvInfo,
}

// RustSBI will now use the `some_other_name` field implementing `rustsbi::Fence`
// as the implementation of SBI Remote Fence extension.

#[derive(RustSBI)]
struct MySBI {
    #[rustsbi(ipi, timer)]
    clint: Clint, // <-- RISC-V CLINT will provide both Ipi and Timer extensions.
    info: MyEnvInfo,
}

// Both Ipi and Timer extension implementations are now provided by the
// `clint` field implementing both `rustsbi::Ipi` and `rustsbi::Timer`.

RustSBI implementations usually provide regular structs to the derive-macro. Alternatively, the RustSBI derive macro also accepts tuple structs or unit structs.

// Tuple structs.
// No field names are provided; the structure must provide helper attributes
// to identify the extensions for the RustSBI derive macro.
#[derive(RustSBI)]
struct MySBI(#[rustsbi(fence)] MyFence, #[rustsbi(info)] MyEnvInfo);

// Unit structs.
// Note that `info` is required in non-machine environment; thus this crate
// only allows unit structs with `machine` feature. No extensions except Base
// extension are provided on unit struct implementations.
#[derive(RustSBI)]
struct MySBI;

Notes

This macro should be used in rustsbi crate as rustsbi::RustSBI.