rustsbi
Struct Forward

pub struct Forward;

Forwards SBI calls onto current supervisor environment.

If crate feature forward is enabled, this structure implements all RustSBI extensions by forwarding the calls into the current supervisor environment. This is done by sbi-rt crate; thus Forward is only available when it’s running in RISC-V SBI environments.

Forward implements all RustSBI traits, but is only effective if #[cfg(feature = "forward")] is enabled. Otherwise, struct Forward is unimplemented!() on SBI calls.

Examples

This structure can be used as a structure field in #[derive(RustSBI)], with helper macro #[rustsbi(extension_1, extension_2, ...)] annotating what extensions should be forwarded to sbi-rt in the structure.

use rustsbi::{Forward, RustSBI};

// Forwards fence, timer and console extensions, but the hsm extension
// is still handled by the `hsm` field variable.
#[derive(RustSBI)]
struct VmSBI {
    hsm: VmHsm,
    #[rustsbi(fence, timer, console, info)]
    forward: Forward,
}